Split status types and checker into different files
